Refactor terminus_judge prep into single script and drop generated data
