+++ resources_servers/terminus_judge/scripts/README.md
@@ -1,41 +1,27 @@
 # Terminus Judge HF Data Pipeline
 
-This directory contains a self-contained 3-stage pipeline for building
-`terminus_judge` training/eval data from HF trajectory datasets.
+Use a single entrypoint:
 
-## Files
-
-- `convert_hf_traj_to_samples.py`: Stage 1 conversion from trajectories to per-turn samples.
-- `make_train_validation_stratified.py`: Stage 2 deduplicated, stratified split builder.
-- `run_smoke_rollouts.sh`: Stage 3 local rollout smoke runner (`ng_run` + `ng_collect_rollouts`).
+- `prepare.py` for conversion, split, and smoke rollout collection.
 
 ## Prerequisites
 
-1. Run with a Python environment that has:
+1. Python env with:
    - `datasets`
    - `openapi-schema-validator`
-2. If your default HF cache path is read-only, set writable cache env vars:
+2. For smoke stage:
+   - `ng_run`, `ng_status`, `ng_collect_rollouts` on `PATH`
+   - reachable policy/model endpoint
+3. If HF cache is read-only:
    - `HF_HOME=/tmp/hf_home`
    - `HF_DATASETS_CACHE=/tmp/hf_home/datasets`
-3. For Stage 3, ensure `ng_run`, `ng_status`, and `ng_collect_rollouts` are available on `PATH`.
-4. Stage 3 also requires a reachable policy model endpoint (or equivalent model-server config).
-5. If Ray startup fails with `pthread_create ... Resource temporarily unavailable`,
-   set low-thread Ray env vars before Stage 3:
-   - `RAY_num_server_call_thread=1`
-   - `RAY_num_grpc_internal_threads=1`
-   - `RAY_enable_worker_prestart=0`
-   - `RAY_worker_maximum_startup_concurrency=1`
-   - `RAY_gcs_server_rpc_server_thread_num=1`
-   - `RAY_gcs_server_rpc_client_thread_num=1`
 
-## Stage 1: Convert Trajectories To Samples
+## Stage 1: Convert HF Trajectories to Samples
 
-Example using local parquet:
+Local parquet:
 
 ```bash
-HF_HOME=/tmp/hf_home \
-HF_DATASETS_CACHE=/tmp/hf_home/datasets \
-python resources_servers/terminus_judge/scripts/convert_hf_traj_to_samples.py \
+python resources_servers/terminus_judge/scripts/prepare.py convert \
   --hf_parquet_glob "datasets/openthoughts/data/*.parquet" \
   --split train \
   --dataset_name openthoughts_agent_v1_sft \
@@ -44,12 +30,10 @@ python resources_servers/terminus_judge/scripts/convert_hf_traj_to_samples.py \
   --threshold 0.95
 ```
 
-Example using HF Hub streaming:
+HF Hub streaming:
 
 ```bash
-HF_HOME=/tmp/hf_home \
-HF_DATASETS_CACHE=/tmp/hf_home/datasets \
-python resources_servers/terminus_judge/scripts/convert_hf_traj_to_samples.py \
+python resources_servers/terminus_judge/scripts/prepare.py convert \
   --hf_dataset open-thoughts/OpenThoughts-Agent-v1-SFT \
   --split train \
   --dataset_name openthoughts_agent_v1_sft \
@@ -58,14 +42,10 @@ python resources_servers/terminus_judge/scripts/convert_hf_traj_to_samples.py \
   --threshold 0.95
 ```
 
-Output:
-
-- `resources_servers/terminus_judge/data/openthoughts_agent_v1_sft/samples.jsonl`
-
-## Stage 2: Build Deduplicated Stratified Train/Validation Splits
+## Stage 2: Deduplicated Stratified Split
 
 ```bash
-python resources_servers/terminus_judge/scripts/make_train_validation_stratified.py \
+python resources_servers/terminus_judge/scripts/prepare.py split \
   --input resources_servers/terminus_judge/data/openthoughts_agent_v1_sft/samples.jsonl \
   --output_dir resources_servers/terminus_judge/data/openthoughts_agent_v1_sft \
   --train_size 0 \
@@ -74,37 +54,20 @@ python resources_servers/terminus_judge/scripts/make_train_validation_stratified
   --seed 42
 ```
 
-Outputs:
-
-- `resources_servers/terminus_judge/data/openthoughts_agent_v1_sft/train.jsonl`
-- `resources_servers/terminus_judge/data/openthoughts_agent_v1_sft/validation.jsonl`
-
 ## Stage 3: Smoke Rollout Collection
 
+By default this uses the canonical 5-row examples already committed in:
+
+- `resources_servers/terminus_judge/data/example.jsonl`
+- `resources_servers/terminus_judge/data/example_rollouts.jsonl`
+
 ```bash
-POLICY_BASE_URL="http://<model-host>:<port>/v1" \
-POLICY_API_KEY="<key>" \
-POLICY_MODEL_NAME="<model-name>" \
-bash resources_servers/terminus_judge/scripts/run_smoke_rollouts.sh
+python resources_servers/terminus_judge/scripts/prepare.py smoke \
+  --policy_base_url "http://<model-host>:<port>/v1" \
+  --policy_api_key "<key>" \
+  --policy_model_name "<model-name>"
 ```
-
-By default this:
-
-1. Copies first `SMOKE_LIMIT` rows from validation to smoke input.
-2. Starts `ng_run`.
-3. Waits for 3 healthy services in `ng_status`.
-4. Runs `ng_collect_rollouts`.
-
-Outputs:
-
-- `resources_servers/terminus_judge/data/openthoughts_agent_v1_sft/smoke/smoke_input.jsonl`
-- `resources_servers/terminus_judge/data/openthoughts_agent_v1_sft/smoke/smoke_rollouts.jsonl`
 
 ## Notes
 
-- The provided `dev/Gym/AGENTS.md` ends at `Sampling:` in Stage 2. For this implementation,
-  Stage 2 uses:
-  - fixed-per-bucket validation sampling (`--val_per_bucket`)
-  - remaining rows for train when `--train_size 0`
-  - otherwise proportional stratified sampling across buckets
-- All required reference files listed in `AGENTS.md` were present and inspected.
+- Generated dataset artifacts (`samples.jsonl`, `train.jsonl`, `validation.jsonl`) are local outputs and are not meant to be committed.
